+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add `Data.Bytes.Parser.Leb128` module for LEB-128 encoding.
   Supports signed integers with zig-zag encoding.
+* Add `skipWhile` to `Data.Bytes.Parser.Latin`.
+* Reexport `endOfInput` and `isEndOfInput` from `Latin`.
 
 ## 0.3.6.0 -- 2020-03-04